# docs/openapi/observability-v1.yaml
openapi: 3.0.3
info:
  title: NeuralOps Observability API
  version: 1.0.0
paths:
...
  /api/v1/notebooks/{id}/execute:
    post:
      summary: Execute notebook cells
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: success envelope
...
  /api/v1/admin/oncall/{id}/sync-pagerduty:
    post:
      summary: Sync rotation from PagerDuty
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: success envelope
...